# Refactor by float conversion steps to macro and apply to fsum().