* Clean-up code by moving float conversion steps into a macro.
* Apply the macro to *fsum()* giving a 20% speed-up for float inputs and a 40% speed-up for integer inputs.

Baseline timings:
```
$ ./python.exe -m timeit -r 11 -s 'from math import fsum' -s 'seq=list(map(float, range(100)))' 'fsum(seq)'
200000 loops, best of 11: 1.27 usec per loop
./python.exe -m timeit -r 11 -s 'from math import fsum' -s 'seq=list(map(int, range(100)))' 'fsum(seq)'
100000 loops, best of 11: 2.26 usec per loop
```

Patched timings:
```
$ ./python.exe -m timeit -r 11 -s 'from math import fsum' -s 'seq=list(map(float, range(100)))' 'fsum(seq)'
200000 loops, best of 11: 1 usec per loop
$ ./python.exe -m timeit -r 11 -s 'from math import fsum' -s 'seq=list(map(int, range(100)))' 'fsum(seq)'
200000 loops, best of 11: 1.36 usec per loop
```